fix: stock padre muestra decimales reales + compras acepta cantidades decimales
- Backend: corrige int() -> float() en available_qty que truncaba 4.995 a 4
- Frontend inventario: padres muestran total_stock (kg reales) no available_qty
- Frontend inventario: muestra hasta 3 decimales significativos con barra visual
- Frontend compras: parseFloat() en lugar de parseInt() para cantidades decimales (ej: 5.5 kg)

+++ backend/app/api/products.py
@@ -467,7 +467,7 @@ def list_products(
             available_from_bom = 0.0
             
         # El available qty es la suma del stock propio (por ejemplo devoluciones/cajas sueltas) + lo que se puede hacer
-        available_qty = float(total_stock) + int(available_from_bom)
+        available_qty = float(total_stock) + float(available_from_bom)
 
         results.append(ProductAggregatedResponse(
             id=primary.id,
